requirements.txt
Django 4.0.4
PyJWT 2.4.0
django-rest-framework 0.1.0
djangorestframework 3.13.1
pip 21.3.1

username ='localuser'
password='password1'